app/requirements.txt
# Core
Django==3.2.13
psycopg2==2.9.3
celery==5.2.7

# Rest apis
djangorestframework==3.13.1
django-cors-headers==3.11.0
drf-spectacular==0.17.3

# Testing
freezegun==1.2.1
model-bakery==1.5.0
requests-mock==1.9.3

# Settings
django-constance[database]==2.8.0

# Other
requests==2.31.0
django-filter==21.1
tenacity==7.0.0
django-health-check==3.16.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field<6.0

# Authentication
datapunt-keycloak-oidc @ git+https://github.com/remyvdwereld/keycloak_oidc_top.git@v1.0
djangorestframework-simplejwt==4.7.2

# Algorithm toolbox
geopy==2.1.0
joblib==1.3.2

#redis
redis==3.5.3

#Azure
azure-identity
applicationinsights==0.11.10
opencensus==0.11.2
opencensus-context==0.1.3
opencensus-ext-azure==1.1.9
opencensus-ext-django==0.8.0
opencensus-ext-logging==0.1.1
opencensus-ext-postgresql==0.1.3
opencensus-ext-requests==0.8.0
